lib/string/strspn/, lib/, src/: Move *spn() APIs to separate subdir
This provides a better organization.
