+++ lib/limits.c
@@ -36,8 +36,8 @@
 #include "atoi/str2i/str2s.h"
 #include "atoi/str2i/str2u.h"
 #include "string/memset/memzero.h"
-#include "string/strchr/stpspn.h"
 #include "string/strcmp/streq.h"
+#include "string/strspn/stpspn.h"
 #include "typetraits.h"
