+++ lib/string/strspn/stpspn.h
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 
-#ifndef SHADOW_INCLUDE_LIB_STRING_STRCHR_STPSPN_H_
-#define SHADOW_INCLUDE_LIB_STRING_STRCHR_STPSPN_H_
+#ifndef SHADOW_INCLUDE_LIB_STRING_STRSPN_STPSPN_H_
+#define SHADOW_INCLUDE_LIB_STRING_STRSPN_STPSPN_H_
 
 
 #include <config.h>
