+++ lib/string/strspn/strrcspn.h
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 
-#ifndef SHADOW_INCLUDE_LIB_STRING_STRCHR_STRRCSPN_H_
-#define SHADOW_INCLUDE_LIB_STRING_STRCHR_STRRCSPN_H_
+#ifndef SHADOW_INCLUDE_LIB_STRING_STRSPN_STRRCSPN_H_
+#define SHADOW_INCLUDE_LIB_STRING_STRSPN_STRRCSPN_H_
 
 
 #include <config.h>
