+++ lib/string/strspn/strrspn.h
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 
-#ifndef SHADOW_INCLUDE_LIB_STRING_STRCHR_STRRSPN_H_
-#define SHADOW_INCLUDE_LIB_STRING_STRCHR_STRRSPN_H_
+#ifndef SHADOW_INCLUDE_LIB_STRING_STRSPN_STRRSPN_H_
+#define SHADOW_INCLUDE_LIB_STRING_STRSPN_STRRSPN_H_
 
 
 #include <config.h>
