+++ lib/strtoday.c
@@ -14,8 +14,8 @@
 #include "atoi/str2i/str2s.h"
 #include "getdate.h"
 #include "prototypes.h"
-#include "string/strchr/stpspn.h"
 #include "string/strcmp/streq.h"
+#include "string/strspn/stpspn.h"
 
 
 /*
