+++ src/login_nopam.c
@@ -59,9 +59,9 @@
 
 #include "prototypes.h"
 #include "sizeof.h"
-#include "string/strchr/strrspn.h"
 #include "string/strcmp/strcaseeq.h"
 #include "string/strcmp/streq.h"
+#include "string/strspn/strrspn.h"
 #include "string/strtok/stpsep.h"
